+++ src/VaultSharp/V1/SecretsEngines/ISecretsEngine.cs
@@ -8,6 +8,7 @@ using VaultSharp.V1.SecretsEngines.Database;
 using VaultSharp.V1.SecretsEngines.GoogleCloud;
 using VaultSharp.V1.SecretsEngines.GoogleCloudKMS;
 using VaultSharp.V1.SecretsEngines.Identity;
+using VaultSharp.V1.SecretsEngines.KMIP;
 using VaultSharp.V1.SecretsEngines.KeyValue;
 using VaultSharp.V1.SecretsEngines.Nomad;
 using VaultSharp.V1.SecretsEngines.PKI;
@@ -67,6 +68,11 @@ namespace VaultSharp.V1.SecretsEngines
         /// The GoogleCloud KMS Secrets Engine.
         /// </summary>
         IGoogleCloudKMSSecretsEngine GoogleCloudKMS { get; }
+
+        /// <summary>
+        /// The KMIP Secrets Engine.
+        /// </summary>
+        IKMIPSecretsEngine KMIP { get; }
 
         /// <summary>
         /// The KeyValue Secrets Engine.