+++ src/VaultSharp/V1/SecretsEngines/SecretsEngineDefaultPaths.cs
@@ -13,6 +13,7 @@
         public const string Database = "database";
         public const string GoogleCloud = "gcp";
         public const string GoogleCloudKMS = "gcpkms";
+        public const string KMIP = "kmip";
         public const string KeyValueV1 = "kv";
         public const string KeyValueV2 = "secret";
         public const string Identity = "identity";